backend/requirements.txt
anthropic==0.69.0
fastapi==0.118.0
pandas==2.3.3
python-dotenv==1.1.1
Requests==2.32.5
yfinance==0.2.66
protobuf==5.26.1
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
uvicorn==0.24.0
slowapi==0.1.9
google-generativeai>=0.3.0
PyJWT==2.8.0
aiohttp==3.9.1
APScheduler==3.10.4

# ML модуль (AI Калькулятор)
torch>=2.0.0
scipy>=1.10.0
numpy>=1.24.0
